Revert "Start to switch to combine"
We want to robust in the face of out of memory errors.
We can't reasonably do that with most of combine's parser combinators....
We're going to rely on 'bumpalo' and arena allocation.

Plus the benefit of the (otherwise awesome) error reporting is minimal,
because Python only raises a 'SyntaxError' for the first error
encountered.

This reverts commit cb7d76cf252da7cc58f45e8dd8c2247cda0aa8ac.